Complete re-write of how the BWT occurrence array markers are represented.
We are using a two-tier system where LargeMarkers are placed every 2048 symbols with the absolute count of the number of symbols seen up to that point.
Every 128 symbols a SmallMarker is placed which holds the count of symbols over the last 128 symbols. This allows us to store these relative counts with
an 8 bit integer instead of a 64 bit for the absolute counters. The absolute counts every 128 symbols are interpolated from the relative counts.
This is a much more space efficient representation - thanks to Travis Wheeler for this suggestion.

This version of the code has the old marker system (absolute counts every 128 symbols) left in and a testing function
is placed at the end of initializeFMIndex(). This version should only be used for testing/debugging the
two-tier system. The old system will be removed in the next revision.