Fixed very subtle bug in overlap computation when reads have multiple valid overlaps to each other.
If, for a given read x, more than two reads have multiple valid overlaps to x it was possible that
when post-processing the overlap blocks to remove the redundant overlaps the forward and reverse
intervals would become out of sync (they would represent different reads). This can effect
the downstream calculations (like the irreducible overlap computation) as the wrong
extensions for one of the reads would be found. This case is very rare so the fix is
to calculate the source read for every element in the overlap block, then remove
the redundant blocks. This calculation is slow but gives the correct result. It has a
minor (a few percent) impact on the total running time. The bug had very little impact
on the final quality of the assembly but the structure of the graph in repetitive
regions is likely more robust now.