Improved the speed of sga-qc by an order of magnitude.
Previously, the interval for each k-mer was computed individually, requiring many Occ() calls. Now, if the interval for a
given kmer k_i meets the threshold, the interval is extended by one base (finding the interval for a k+1-mer).
If this k+1 interval meets the threshold, it implies that k_(i+1) also meets the threshold which saves k - 1 calls to Occ().