Refined the kmer matching portion of the overlap calculation for the new corrector.
Previously, we called SampledSuffixArray::calcSA to get the read ID for every kmer match.
This was very expensive as usually reads would share multiple kmers, making the calculation
very redundant. Now we cache the indices as we visit them which is two orders of magnitude
faster. The DP overlap calculation is now the bottleneck.